services/feature-builder/pyproject.toml
[tool.poetry]
name = "feature-builder"
version = "0.1.0"
description = "CS2 Feature Builder service: compute and cache features for teams/players/matches"
authors = ["faztyG <verselabproject@gmail.com>"]
packages = [{ include = "feature_builder" }]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
fastapi = "^0.112.0"
uvicorn = {version = "^0.30.0", extras = ["standard"]}
redis = "^5.0.0"
nats-py = "^2.11.0"
orjson = "^3.9.0"
httpx = "^0.28.0"
asyncpg = "^0.29.0"
pydantic-settings = "^2.4.0"
pandas = "^2.2.0"
numpy = "^2.0.0"
scipy = "^1.12.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
